fix int not the right size on 64 bit systems
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4141 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/logic/worker.cc
@@ -800,7 +800,7 @@ bool Worker::run_geologist_find(Game & game, State & state, Action const &)
 		 */
 		if (rdescr && rdescr->is_detectable()) {
 			log("Message: resources found %s\n", rdescr->name().c_str());
-			std::string sender = "g " + reinterpret_cast<int>(this);
+			std::string sender = "g " + reinterpret_cast<intptr_t>(this);
 			MessageQueue::add
 				(owner(),
 				 Message
